move virtual4c into hist.hicfeature module; add doc in hist.hicfeature
