+++ coolbox/core/track/__init__.py
@@ -4,9 +4,8 @@ from .bed import BED
 from .gtf import GTF
 from .hicdiff import HiCDiff, Selfish
 from .hicmat import HiCMat, Cool, DotHiC
-from .hist import Hist, BigWig, ABCompartment, BedGraph, DiScore, InsuScore
+from .hist import Hist, BigWig, ABCompartment, BedGraph, DiScore, InsuScore, Virtual4C
 from .ideogram import Ideogram
 from .pseudo import Spacer, HLine, XAxis, ChromName
 from .snp import SNP
 from .tads import TADs
-from .v4c import Virtual4C
