+++ coolbox/core/track/hicdiff/selfish.py
@@ -1,3 +1,4 @@
+from abc import ABC
 import numpy as np
 import pandas as pd
 import statsmodels.stats.multitest as smm
