+++ coolbox/core/track/hicmat/__init__.py
@@ -1,13 +1,22 @@
-from coolbox.utilities.hic.tools import hicmat_filetype
+from pathlib import Path
+from typing import Union
 from .base import HicMatBase
 from .cool import Cool
 from .dothic import DotHiC
 
 
-def HiCMat(file_, *args, **kwargs) -> HicMatBase:
+def HiCMat(file_or_hicmat: Union[str, HicMatBase], *args, **kwargs) -> HicMatBase:
     """Compose Hi-C track(.cool, .mcool, .hic), determine type by file extension."""
-    ftype = hicmat_filetype(file_)
-    if ftype == ".hic":
-        return DotHiC(file_, *args, **kwargs)
+    if isinstance(file_or_hicmat, HicMatBase):
+        return file_or_hicmat
+    elif not Path(file_or_hicmat).is_file():
+        raise ValueError("The file path does not exist.")
+
+    p = file_or_hicmat
+    if p.endswith(".hic"):
+        return DotHiC(file_or_hicmat, *args, **kwargs)
+    elif p.endswith(".cool") or p.endswith(".mcool"):
+        return Cool(file_or_hicmat, *args, **kwargs)
     else:
-        return Cool(file_, *args, **kwargs)
+        raise NotImplementedError(f"File type of {p} not supported for HicMat. "
+                                  f"The file type should be one of .cool/.mcool/.hic")
