+++ coolbox/core/track/hist/__init__.py
@@ -1,14 +1,18 @@
 from .bedgraph import BedGraph
 from .bigwig import BigWig, ABCompartment
-from .hicfeatures import DiScore, InsuScore
+from .hicfeature import DiScore, InsuScore, Virtual4C
+from .base import HistBase
 
 
-def Hist(file, **kwargs):
+def Hist(file, *args, **kwargs) -> HistBase:
     """Hist track(bigwig, bedgraph), determine track type by file extension."""
-    if file.endswith(".bw") or file.endswith(".bigwig"):
-        return BigWig(file, **kwargs)
-    elif file.endswith('.bedgraph') or file.endswith('.bg') or \
-            file.endswith('.bedgraph.bgz') or file.endswith('.bg.bgz'):
-        return BedGraph(file, **kwargs)
+    if file.endswith(".bw") or \
+            file.endswith(".bigwig"):
+        return BigWig(file, *args, **kwargs)
+    elif file.endswith('.bedgraph') or \
+            file.endswith('.bg') or \
+            file.endswith('.bedgraph.bgz') or \
+            file.endswith('.bg.bgz'):
+        return BedGraph(file, *args, **kwargs)
     else:
         raise NotImplementedError("Hist only support .bigwig or .bedgraph file now.")
